+++ dds_web/utils.py
@@ -15,7 +15,6 @@ from contextlib import contextmanager
 import flask
 import sqlalchemy
 import pytz
-import marshmallow
 
 # # imports related to scheduling
 import atexit
